Fixed and streamlined serializaling statements on storage operations
* Properly quoted column identifiers when copying a table
* Reused `streaming_*()` functionality
