+++ dev/implementations/storage_definitions.h
@@ -8,9 +8,10 @@
 #include <functional>  //  std::reference_wrapper, std::cref
 #include <algorithm>  //  std::find_if
 
-#include "../storage.h"
 #include "../dbstat.h"
 #include "../util.h"
+#include "../serializing_util.h"
+#include "../storage.h"
 
 namespace sqlite_orm {
     namespace internal {
@@ -96,36 +97,24 @@ namespace sqlite_orm {
                                           const std::string& tableName,
                                           const I& tImpl,
                                           const std::vector<table_info*>& columnsToIgnore) const {
-            std::stringstream ss;
-            std::vector<std::string> columnNames;
-            tImpl.table.for_each_column([&columnNames, &columnsToIgnore](auto& c) {
+            std::vector<std::reference_wrapper<const std::string>> columnNames;
+            columnNames.reserve(tImpl.table.count_columns_amount());
+            tImpl.table.for_each_column([&columnNames, &columnsToIgnore](const auto& c) {
                 auto& columnName = c.name;
-                auto columnToIgnoreIt =
-                    std::find_if(columnsToIgnore.begin(), columnsToIgnore.end(), [&columnName](auto tableInfoPointer) {
-                        return columnName == tableInfoPointer->name;
-                    });
+                auto columnToIgnoreIt = std::find_if(columnsToIgnore.begin(),
+                                                     columnsToIgnore.end(),
+                                                     [&columnName](const table_info* tableInfo) {
+                                                         return columnName == tableInfo->name;
+                                                     });
                 if(columnToIgnoreIt == columnsToIgnore.end()) {
-                    columnNames.emplace_back(columnName);
+                    columnNames.push_back(cref(columnName));
                 }
             });
-            auto columnNamesCount = columnNames.size();
-            ss << "INSERT INTO " << quote_identifier(tableName) << " (";
-            for(size_t i = 0; i < columnNamesCount; ++i) {
-                ss << columnNames[i];
-                if(i < columnNamesCount - 1) {
-                    ss << ",";
-                }
-                ss << " ";
-            }
-            ss << ") ";
-            ss << "SELECT ";
-            for(size_t i = 0; i < columnNamesCount; ++i) {
-                ss << columnNames[i];
-                if(i < columnNamesCount - 1) {
-                    ss << ", ";
-                }
-            }
-            ss << " FROM " << quote_identifier(tImpl.table.name);
+
+            std::stringstream ss;
+            ss << "INSERT INTO " << streaming_identifier(tableName) << " (" << streaming_identifiers(columnNames)
+               << ") "
+               << "SELECT " << streaming_identifiers(columnNames) << " FROM " << streaming_identifier(tImpl.table.name);
             perform_void_exec(db, ss.str());
         }
