+++ dev/statement_serializer.h
@@ -36,6 +36,7 @@
 #include "table_name_collector.h"
 #include "column_names_getter.h"
 #include "order_by_serializer.h"
+#include "serializing_util.h"
 #include "statement_binder.h"
 #include "values.h"
 #include "triggers.h"
@@ -54,322 +55,6 @@ namespace sqlite_orm {
         std::string serialize(const T& t, const serializer_context<I>& context) {
             statement_serializer<T> serializer;
             return serializer(t, context);
-        }
-
-#if __cplusplus >= 202002L  // C++20 or later
-        inline void stream_sql_escaped(std::ostream& os, const std::string& str, char char2Escape) {
-            for(std::string::const_iterator it = str.cbegin(), next; true; it = next + 1) {
-                next = std::find(it, str.cend(), char2Escape);
-                os << std::string_view{it, next};
-
-                if(next == str.cend()) [[likely]] {
-                    break;
-                }
-                os << std::string(2, char2Escape);
-            }
-        }
-#else
-        inline void stream_sql_escaped(std::ostream& os, const std::string& str, char char2Escape) {
-            if(str.find(char2Escape) == str.npos) {
-                os << str;
-            } else {
-                for(char c: str) {
-                    if(c == char2Escape) {
-                        os << char2Escape;
-                    }
-                    os << c;
-                }
-            }
-        }
-#endif
-
-        inline void stream_identifier(std::ostream& ss,
-                                      const std::string& qualifier,
-                                      const std::string& identifier,
-                                      const std::string& alias) {
-            constexpr char quoteChar = '"';
-            constexpr char qualified[] = {quoteChar, '.', '\0'};
-            constexpr char aliased[] = {' ', quoteChar, '\0'};
-
-            // note: In practice, escaping double quotes in identifiers is arguably overkill,
-            // but since the SQLite grammar allows it, it's better to be safe than sorry.
-
-            if(!qualifier.empty()) {
-                ss << quoteChar;
-                stream_sql_escaped(ss, qualifier, quoteChar);
-                ss << qualified;
-            }
-            {
-                ss << quoteChar;
-                stream_sql_escaped(ss, identifier, quoteChar);
-                ss << quoteChar;
-            }
-            if(!alias.empty()) {
-                ss << aliased;
-                stream_sql_escaped(ss, alias, quoteChar);
-                ss << quoteChar;
-            }
-        }
-
-        inline void stream_identifier(std::ostream& ss, const std::string& identifier, const std::string& alias) {
-            return stream_identifier(ss, std::string{}, identifier, alias);
-        }
-
-        inline void stream_identifier(std::ostream& ss, const std::string& identifier) {
-            return stream_identifier(ss, std::string{}, identifier, std::string{});
-        }
-
-        template<typename Tpl, size_t... Is>
-        void stream_identifier(std::ostream& ss, const Tpl& tpl, std::index_sequence<Is...>) {
-            return stream_identifier(ss, std::get<Is>(tpl)...);
-        }
-
-        template<typename Tpl>
-        void stream_identifier(std::ostream& ss, const Tpl& tpl) {
-            return stream_identifier(ss, tpl, std::make_index_sequence<std::tuple_size<Tpl>::value>{});
-        }
-
-        enum class stream_as {
-            conditions_tuple,
-            actions_tuple,
-            expressions_tuple,
-            dynamic_expressions,
-            serialized,
-            identifier,
-            identifiers,
-            values_placeholders,
-            table_columns,
-            non_generated_columns,
-            mapped_columns_expressions,
-        };
-
-        template<stream_as mode>
-        struct streaming {
-            template<class... Ts>
-            auto operator()(const Ts&... ts) const {
-                return std::forward_as_tuple(*this, ts...);
-            }
-
-            template<size_t... Idx>
-            constexpr std::index_sequence<1u + Idx...> offset_index(std::index_sequence<Idx...>) const {
-                return {};
-            }
-        };
-        SQLITE_ORM_INLINE_VAR constexpr streaming<stream_as::conditions_tuple> streaming_conditions_tuple{};
-        SQLITE_ORM_INLINE_VAR constexpr streaming<stream_as::actions_tuple> streaming_actions_tuple{};
-        SQLITE_ORM_INLINE_VAR constexpr streaming<stream_as::expressions_tuple> streaming_expressions_tuple{};
-        SQLITE_ORM_INLINE_VAR constexpr streaming<stream_as::dynamic_expressions> streaming_dynamic_expressions{};
-        SQLITE_ORM_INLINE_VAR constexpr streaming<stream_as::serialized> streaming_serialized{};
-        SQLITE_ORM_INLINE_VAR constexpr streaming<stream_as::identifier> streaming_identifier{};
-        SQLITE_ORM_INLINE_VAR constexpr streaming<stream_as::identifiers> streaming_identifiers{};
-        SQLITE_ORM_INLINE_VAR constexpr streaming<stream_as::values_placeholders> streaming_values_placeholders{};
-        SQLITE_ORM_INLINE_VAR constexpr streaming<stream_as::table_columns> streaming_table_column_names{};
-        SQLITE_ORM_INLINE_VAR constexpr streaming<stream_as::non_generated_columns>
-            streaming_non_generated_column_names{};
-        SQLITE_ORM_INLINE_VAR constexpr streaming<stream_as::mapped_columns_expressions>
-            streaming_mapped_columns_expressions{};
-
-        // serialize and stream a tuple of condition expressions;
-        // space + space-separated
-        template<class T, class Ctx>
-        std::ostream& operator<<(std::ostream& ss, std::tuple<decltype((streaming_conditions_tuple)), T, Ctx> tpl) {
-            const auto& conditions = get<1>(tpl);
-            auto& context = get<2>(tpl);
-
-            iterate_tuple(conditions, [&ss, &context](auto& c) {
-                ss << " " << serialize(c, context);
-            });
-            return ss;
-        }
-
-        // serialize and stream a tuple of action expressions;
-        // space-separated
-        template<class T, class Ctx>
-        std::ostream& operator<<(std::ostream& ss, std::tuple<decltype((streaming_actions_tuple)), T, Ctx> tpl) {
-            const auto& actions = get<1>(tpl);
-            auto& context = get<2>(tpl);
-
-            bool first = true;
-            iterate_tuple(actions, [&ss, &context, &first](auto& a) {
-                constexpr std::array<const char*, 2> sep = {" ", ""};
-                ss << sep[std::exchange(first, false)] << serialize(a, context);
-            });
-            return ss;
-        }
-
-        // serialize and stream a tuple of expressions;
-        // comma-separated
-        template<class T, class Ctx>
-        std::ostream& operator<<(std::ostream& ss, std::tuple<decltype((streaming_expressions_tuple)), T, Ctx> tpl) {
-            const auto& args = get<1>(tpl);
-            auto& context = get<2>(tpl);
-
-            bool first = true;
-            iterate_tuple(args, [&ss, &context, &first](auto& arg) {
-                constexpr std::array<const char*, 2> sep = {", ", ""};
-                ss << sep[std::exchange(first, false)] << serialize(arg, context);
-            });
-            return ss;
-        }
-
-        // serialize and stream multi_order_by arguments;
-        // comma-separated
-        template<class... Os, class Ctx>
-        std::ostream&
-        operator<<(std::ostream& ss,
-                   std::tuple<decltype((streaming_expressions_tuple)), const std::tuple<order_by_t<Os>...>&, Ctx> tpl) {
-            const auto& args = get<1>(tpl);
-            auto& context = get<2>(tpl);
-
-            bool first = true;
-            iterate_tuple(args, [&ss, &context, &first](auto& arg) {
-                constexpr std::array<const char*, 2> sep = {", ", ""};
-                ss << sep[std::exchange(first, false)] << serialize_order_by(arg, context);
-            });
-            return ss;
-        }
-
-        // serialize and stream a vector of expressions;
-        // comma-separated
-        template<class C, class Ctx>
-        std::ostream& operator<<(std::ostream& ss, std::tuple<decltype((streaming_dynamic_expressions)), C, Ctx> tpl) {
-            const auto& args = get<1>(tpl);
-            auto& context = get<2>(tpl);
-
-            constexpr std::array<const char*, 2> sep = {", ", ""};
-            for(size_t i = 0, first = true; i < args.size(); ++i) {
-                ss << sep[std::exchange(first, false)] << serialize(args[i], context);
-            }
-            return ss;
-        }
-
-        // stream a vector of already serialized strings;
-        // comma-separated
-        template<class C>
-        std::ostream& operator<<(std::ostream& ss, std::tuple<decltype((streaming_serialized)), C> tpl) {
-            const auto& strings = get<1>(tpl);
-
-            constexpr std::array<const char*, 2> sep = {", ", ""};
-            for(size_t i = 0, first = true; i < strings.size(); ++i) {
-                ss << sep[std::exchange(first, false)] << strings[i];
-            }
-            return ss;
-        }
-
-        // stream an identifier described by a variadic string pack, which is one of:
-        // 1. identifier
-        // 2. identifier, alias
-        // 3. qualifier, identifier, alias
-        template<class... Strings>
-        std::ostream& operator<<(std::ostream& ss, std::tuple<decltype((streaming_identifier)), Strings...> tpl) {
-            stream_identifier(ss, tpl, streaming_identifier.offset_index(std::index_sequence_for<Strings...>{}));
-            return ss;
-        }
-
-        // stream a container of identifiers described by a string or a tuple, which is one of:
-        // 1. identifier
-        // 1. tuple(identifier)
-        // 2. tuple(identifier, alias), pair(identifier, alias)
-        // 3. tuple(qualifier, identifier, alias)
-        //
-        // comma-separated
-        template<class C>
-        std::ostream& operator<<(std::ostream& ss, std::tuple<decltype((streaming_identifiers)), C> tpl) {
-            const auto& identifiers = get<1>(tpl);
-
-            constexpr std::array<const char*, 2> sep = {", ", ""};
-            bool first = true;
-            for(auto& identifier: identifiers) {
-                ss << sep[std::exchange(first, false)];
-                stream_identifier(ss, identifier);
-            }
-            return ss;
-        }
-
-        // stream placeholders as part of a values clause
-        template<class... Ts>
-        std::ostream& operator<<(std::ostream& ss, std::tuple<decltype((streaming_values_placeholders)), Ts...> tpl) {
-            const size_t& columnsCount = get<1>(tpl);
-            const ptrdiff_t& valuesCount = get<2>(tpl);
-
-            if(!valuesCount || !columnsCount) {
-                return ss;
-            }
-
-            std::string result;
-            result.reserve((1 + (columnsCount * 1) + (columnsCount * 2 - 2) + 1) * valuesCount + (valuesCount * 2 - 2));
-
-            constexpr std::array<const char*, 2> sep = {", ", ""};
-            for(ptrdiff_t i = 0, first = true; i < valuesCount; ++i) {
-                result += sep[std::exchange(first, false)];
-                result += "(";
-                for(size_t i = 0, first = true; i < columnsCount; ++i) {
-                    result += sep[std::exchange(first, false)];
-                    result += "?";
-                }
-                result += ")";
-            }
-            ss << result;
-            return ss;
-        }
-
-        // stream a table's column identifiers, possibly qualified;
-        // comma-separated
-        template<class Table>
-        std::ostream& operator<<(std::ostream& ss,
-                                 std::tuple<decltype((streaming_table_column_names)), Table, const bool&> tpl) {
-            const auto& table = get<1>(tpl);
-            const bool& qualified = get<2>(tpl);
-
-            bool first = true;
-            table.for_each_column([&ss, &tableName = qualified ? table.name : std::string{}, &first](auto& column) {
-                constexpr std::array<const char*, 2> sep = {", ", ""};
-                ss << sep[std::exchange(first, false)];
-                stream_identifier(ss, tableName, column.name, std::string{});
-            });
-            return ss;
-        }
-
-        // stream a table's non-generated column identifiers, unqualified;
-        // comma-separated
-        template<class Table>
-        std::ostream& operator<<(std::ostream& ss,
-                                 std::tuple<decltype((streaming_non_generated_column_names)), Table> tpl) {
-            const auto& table = get<1>(tpl);
-
-            bool first = true;
-            table.for_each_column([&ss, &first](auto& column) {
-                if(column.is_generated()) {
-                    return;
-                }
-
-                constexpr std::array<const char*, 2> sep = {", ", ""};
-                ss << sep[std::exchange(first, false)];
-                stream_identifier(ss, column.name);
-            });
-            return ss;
-        }
-
-        // stream a tuple of mapped columns (which are member pointers or column pointers);
-        // comma-separated
-        template<class T, class Ctx>
-        std::ostream& operator<<(std::ostream& ss,
-                                 std::tuple<decltype((streaming_mapped_columns_expressions)), T, Ctx> tpl) {
-            const auto& columns = get<1>(tpl);
-            auto& context = get<2>(tpl);
-
-            bool first = true;
-            iterate_tuple(columns, [&ss, &context, &first](auto& column) {
-                const std::string* columnName = find_column_name(context.impl, column);
-                if(!columnName) {
-                    throw std::system_error{orm_error_code::column_not_found};
-                }
-
-                constexpr std::array<const char*, 2> sep = {", ", ""};
-                ss << sep[std::exchange(first, false)];
-                stream_identifier(ss, *columnName);
-            });
-            return ss;
         }
 
         /**
@@ -521,7 +206,9 @@ namespace sqlite_orm {
 
             template<class Ctx>
             std::string operator()(const statement_type&, const Ctx&) {
-                return quote_identifier(T::get());
+                std::stringstream ss;
+                ss << streaming_identifier(T::get());
+                return ss.str();
             }
         };
